backend/requirements.txt
# Flask Core
Flask==3.1.0
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.7.1

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# Redis & Celery
redis==5.2.1
celery==5.4.0

# IA & ML
transformers==4.47.1
torch>=2.6.0
huggingface-hub==0.27.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2

# API & HTTP
requests==2.32.3
httpx==0.28.1
flask-restx==1.3.0

# Validation
marshmallow==3.23.2
python-dotenv==1.0.1

# Security
bcrypt==4.2.1
cryptography==44.0.0
Flask-WTF==1.2.1

# Monitoring
prometheus-flask-exporter==0.23.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2

